feat: add `WriteText` action
