+++ packages/fluttium/lib/src/actions/actions.dart
@@ -1,7 +1,7 @@
 export 'action.dart';
 export 'expect_not_visible.dart';
 export 'expect_visible.dart';
-export 'input_text.dart';
 export 'long_press_on.dart';
 export 'press_on.dart';
 export 'take_screenshot.dart';
+export 'write_text.dart';
